+++ client/src/constants/index.jsx
@@ -43,13 +43,10 @@ export {
   GET_EVENTS,
   GET_ALL_TOPICS,
   GET_USER,
-  // SUBCR_USER_TO_EVENT,
-  // UNSUBCR_USER_FROM_EVENT,
   SUBCR_USER_TO_TOPIC,
   UNSUBCR_USER_FROM_TOPIC,
   SET_USER_DEPARTMENT,
   GET_ALL_DEPARTMENTS,
-  // GET_EVENT,
   GET_TOPIC,
   token,
 };
