+++ client/src/pages/LoginPage/index.jsx
@@ -7,28 +7,14 @@ import Header from "../../common/Header";
 import axios from "axios";
 import { SET_USER_DATA } from "../../constants";
 import jwtDecode from "jwt-decode";
-import { setStorage, router } from "./helpers";
+import { setStorage, router } from "../../helpers/helpers";
 import ErrorMessage from "../../components/ErrorMessage";
 import BanMsg from "../../components/BanMsg";
 
 export default class LoginPage extends Component {
-  // state = {
-  //   error: "",
-  // };
-
   render() {
     // const isBanned = sessionStorage.getItem("banned") === "true" ? true : false;
     const handleTelegramResponse = async telegramResponse => {
-      // old api 1.0
-      // const requestObj = {
-      //   method: "put",
-      //   url: SET_USER_DATA,
-      //   data: telegramResponse,
-      //   mode: "cors",
-      //   "Content-Type": "application/json",
-      // };
-      //----------------------------------------
-      // new api 2.0
       const requestObj = {
         method: "post",
         url: "https://forgeserver.herokuapp.com/login/",
@@ -36,23 +22,13 @@ export default class LoginPage extends Component {
         mode: "cors",
         "Content-Type": "application/json",
       };
-      //--------------------------------------
-      console.log("probros 1");
       const token = await axios(requestObj)
         .then(response => {
-          console.log("probros 2");
           return response.data.token;
         })
         .catch(err => {
-          // this.setState(() => {
-          //   return {
-          //     error: err,
-          //   };
-          // });
-          // console.log(err);
           return err;
         });
-      console.log("probros 3");
       const userData = jwtDecode(`${token}`);
       const date = new Date(userData.exp * 1000).toGMTString();
       setCookie("token", token, { expires: date });
