+++ client/src/pages/PageTeamSelect/index.jsx
@@ -23,7 +23,6 @@ const setUserDepartment = departmentId => {
 
   const userId = sessionStorage.getItem("id");
   const token = getCookie("token");
-  // api 2.0
   return axios({
     method: "put",
     url: `${SET_USER_DEPARTMENT(userId)}`,
@@ -35,32 +34,10 @@ const setUserDepartment = departmentId => {
     },
   })
     .then(res => {
-      // const user = jwtdecode(`${res.data.token}`);
-      // console.log(user);
-      // sessionStorage.setItem("department", user.data.department._id);
-      // console.log(res);
       return res;
     })
     .catch(error => console.log(error));
 };
-//---------------------------------------------
-
-// new api 1.0
-//   return axios
-//     .put(
-//       `${SET_USER_DEPARTMENT(userId)}`,
-//       { newDepartment: departmentId },
-//       {
-//         headers: {
-//           Authorization: `Bearer ${getCookie("token")}`,
-//           "Content-Type": "application/json",
-//           mode: "cors",
-//         },
-//       },
-//     )
-//     .catch(error => console.log(error));
-// };
-//--------------------------------------------------------
 
 const PageTeamSelect = props => {
   const [selectedDepartmentId, setSelectedDepartmentId] = useState(null);
@@ -71,29 +48,18 @@ const PageTeamSelect = props => {
     // checkTokenTime(sessionStorage.getItem("tokenTimeOver"));
 
     const fetchData = async () => {
-      // api 1.0
-      // const result = await axios(GET_ALL_DEPARTMENTS, {
-      //   headers: {
-      //     Authorization: `Bearer ${getCookie("token")}`,
-      //   },
-      // });
-      //--------------------------------------------------------
-
-      // new api 2.0
       const result = await axios(GET_ALL_DEPARTMENTS, {
         headers: {
           Authorization: `Bearer ${getCookie("token")}`,
         },
-      });
+      })
+        .then(res => res)
+        .catch(err => {
+          console.log(err);
+          return err;
+        });
       // console.log(result.data.data);
-      //--------------------------------------------------------
-      //api 1.0
-      // setOptions(getAccountOptions(result.data));
-      //--------------------------------------------------------
-
-      //  api 2.0
       setOptions(getAccountOptions(result.data.data));
-      //--------------------------------------------------------
     };
 
     fetchData();
